Custombackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.19.0
alembic==1.12.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
bcrypt==5.0.0
bitarray==3.8.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.2.1
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.128.0
filterpy==1.4.5
fonttools==4.61.0
frozenlist==1.8.0
GeoAlchemy2==0.14.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.8.2
mcp==1.25.0
multidict==6.7.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==25.0
paho-mqtt==2.1.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.9
pyais==2.6.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.21
pytz==2023.3
pywin32==311; sys_platform == 'win32'
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
rsa==4.9.1
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sse-starlette==3.1.2
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==12.0
win32_setctime==1.2.0; sys_platform == 'win32'
yarl==1.22.0